testing int instead of float

+++ src/time4riscv/Entities.c
@@ -12,7 +12,7 @@ void Player_init(volatile Player* p, int id, int color, int x_pos, int y_pos) {
     p->velocity = 1; // Inital velocity
     p->dx = 0;
     p->dy = 0;
-    p->radius = sqrt(p->area/3.14f);
+    p->radius = (int) sqrt(p->area/3.14f);
     p->id = id;
     print("-------- Player initialized\n");
 }
@@ -20,12 +20,12 @@ void Player_init(volatile Player* p, int id, int color, int x_pos, int y_pos) {
 void Ai_init(volatile Ai* ai, int id, int color, int x_pos, int y_pos) {
     ai->x_pos = x_pos; //TODO: Random start position
     ai->y_pos = y_pos; //TODO: Random start position
-    ai->area = 100.0f;
+    ai->area = 100;
     ai->color = color;
     ai->velocity = 1; // Inital velocity
     ai->dx = 0;
     ai->dy = 0;
-    ai->radius = sqrt(ai->area/3.14f);
+    ai->radius = (int) sqrt(ai->area/3.14f);
     ai->id = id;
     print("-------- Calculating radius for AI with area %.2f\n", ai->area);
     print("-------- AI initialized: id=%d, pos=(%d,%d), color=%d, radius=%.2f\n", ai->id, ai->x_pos, ai->y_pos, ai->color, ai->radius);
@@ -119,7 +119,7 @@ void Player_update_position(volatile Player* p, volatile GameState* gs, int x_ct
 
 void Player_update_velocity(volatile Player* p) {
     int base_velocity = 1;
-    p->velocity = base_velocity - ((int)p->area / 100);
+    p->velocity = base_velocity - (int) (p->area / 100);
     if (p->velocity < 1) {
         p->velocity = 1;
     }
@@ -181,7 +181,7 @@ void AI_update_position(volatile Ai* ai, volatile GameState* gs, int x_ctrl, int
 
 void AI_update_velocity(volatile Ai* ai) {
     int base_velocity = 1;
-    ai->velocity = base_velocity - ((int)ai->area / 100);
+    ai->velocity = base_velocity - (int) (ai->area / 100);
     if (ai->velocity < 1) {
         ai->velocity = 1;
     }
